+++ packages/safe-ds-lang/tests/language/documentation/safe-ds-comment-provider.test.ts
@@ -1,9 +1,7 @@
-import { afterEach, describe, expect, it } from 'vitest';
-import { createSafeDsServices } from '../../../src/language/safe-ds-module.js';
+import { AssertionError } from 'assert';
 import { AstNode, EmptyFileSystem } from 'langium';
 import { clearDocuments } from 'langium/test';
-import { getNodeOfType } from '../../helpers/nodeFinder.js';
-import { AssertionError } from 'assert';
+import { afterEach, describe, expect, it } from 'vitest';
 import {
     isSdsAnnotation,
     isSdsAttribute,
@@ -15,6 +13,8 @@ import {
     isSdsResult,
     isSdsTypeParameter,
 } from '../../../src/language/generated/ast.js';
+import { createSafeDsServices } from '../../../src/language/index.js';
+import { getNodeOfType } from '../../helpers/nodeFinder.js';
 
 const services = createSafeDsServices(EmptyFileSystem).SafeDs;
 const commentProvider = services.documentation.CommentProvider;
